Test/requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
lxml==4.9.1
postgres==4.0
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pycparser==2.21
PySocks==1.7.1
requests==2.28.1
selenium==4.4.3
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41

